Add time codes to player
Adjust track component
